+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+- 0.1.0
+
+* New 'Unwrap' class for unpacking
+
+```dart
+String? name = "king";
+Unwrap(name).map((e) => "hello $e"); /// hello king
+```
+
 - 0.0.2
 
 * fix double to int error